pyproject.toml
[tool.poetry]
name = "llwinecellar"
version = "0.1.0"
description = ""
authors = ["lynxlevin <72173230+lynxlevin@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1.1"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.3"
django-environ = "^0.9.0"
mypy = "^0.971"
requests = "^2.28.1"
ruff = "^0.0.265"
gunicorn = "^21.2.0"

[tool.poetry.dev-dependencies]
factory-boy = "^3.2.1"
django-silk = "^5.1.0"

[tool.ruff]
# https://beta.ruff.rs/docs/settings/
select = ["E", "F", "I"]
line-length = 120
unfixable = ["F401", "F841"]
[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"